Version qui marche

+++ Voiture.h
@@ -1,10 +1,12 @@
 #pragma once
 #include "MoyenTransport.h"
 #include <iostream>
+#include <string>
+
 using namespace std;
 class Voiture : public MoyenTransport {
 
 public:
     Voiture();
-    Voiture(string nom, float vitesse);
+    Voiture(string nom, double vitesse);
 } ;
